server/requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
python-dotenv
transformers[torch]
torch
aiohttp
requests
psutil
sentencepiece
